### 2025-07-09 1c57355
trim paper.md
