+++ README.Rmd
@@ -201,13 +201,13 @@ The `gibbs_ad_type()` function takes the following inputs, similar to `gibbs_ad(
 * `id` : A vector of the `id` of one or more specific finds whose use date is to be estimated. The values of `id` must match those in the `list` of `finds`. If `type` is used, `id` is ignored.
 * `type` : A vector of one or more types to estimate a use density for. Must contain a value if `id` is left as `NULL`.
 * `type_name` : A customized label for the type (e.g., if one is pooling together multiple `id`/`type` entries). If only one `type` has been entered, that label is used. Otherwise it defaults to just `"Type"`.
-* `max_samples`, `size`, `mcse_crit`, `trim` : The same information used for determining the maximum length of the Gibbs sampler and when convergence has been achieved, as well as whether to trim events, [as above](#dates-of-events-in-sequences-and-absolute-constraints). Note that the `mcse_crit`, as a stopping rule, applies still to the sequential events/absolute constraints, but MCSE will still be reported for the estimates of the production, use, and deposition dates.
+* `max_samples`, `size`, `mcse_crit`, `trim` : The same information used for determining the maximum length of the Gibbs sampler and when convergence has been achieved, as well as whether to trim events, [as above](#dates-of-events-in-sequences-and-absolute-constraints). Note that the `mcse_crit`, as a stopping rule, applies still to the sequential events/absolute constraints, but MCSE will still be reported for the estimates of the production, use, and depositional dates.
 * `tpq` and `taq` :  Format must follow that illustrated in the Section above on [Absolute Constraints](#absolute-constraints). 
 * `rule`: the rule for determining the earliest date of production of an artifact type. Initial threshold boundaries are first established between the earliest depositional context containing an artifact of that type and the next earliest context which lacks it. Then, the following rules will sample a date accordingly:
   * `naive`: samples are drawn between the initial threshold sample and the depositional date of that artifact.
   * `earliest`: samples are drawn within the initial threshold boundaries.
 
-As use dates are drawn between production and depositional dates, if one chooses `"earliest"` as the rule, then the use density is equivalent to that of the `"naive"` production density. It should also be noted that, for this function, Gibbs sampling is only used for the depositional sequences and absolute constraints, not for production, use, and deposition (i.e, the use date does not affect the production date, nor is the deposition date affected by the production date).
+As use dates are drawn between production and depositional dates, if one chooses `"earliest"` as the rule, then the use density is equivalent to that of the `"naive"` production density. It should also be noted that, for this function, Gibbs sampling is only used for the depositional sequences and absolute constraints, not for production, use, and deposition (i.e, the use date does not affect the production date, nor is the depositional date affected by the production date).
 
 Using the `result` object above, the densities of the use dates of the following types is computed using the `gibbs_ad_type()` function as follows:
 
