### 2025-07-09 cb4dfb6
Update paper.md
