+++ paper/paper.md
@@ -19,13 +19,13 @@ bibliography: ../inst/REFERENCES.bib
 
 # Summary
 
-Archaeological chronology-building centers on two types of dates, absolute and relative. Absolute dates are associated with a calendrical date (even approximate), and relative refer to events only known via relationships to others (before/after). Relative events are then associated to absolute via contextual or logical associations. E.g, a _terminus post quem_ (_t.p.q._) of a datable coin found in a soil deposit ensures the deposit came after the production of the coin. Applying probabilistic methods to chronology has traditionally centered on radiocarbon dating, but the broader goal of extending formal methods to all aspects of chronology necessitate a way to address artifact typologies. Artifacts largely persist in being dated by qualitative judgment (e.g, "around the start of the 4th century BCE"), and their production and use transcend assumptions about depositional sequences (e.g., finds can be made elsewhere and well before a site is occupied). Moreover, typologies are in a constant state of revision and adjustment, such that the dates of types can differ according to authority. The reasons why a type is dated the way it is (i.e., which sites, deposits, and their comparisons inform its date) can also become opaque, such that investigators inherently resist chronological revisions in order to assess the bibliography on any one type (on this conservatism, see @rotroff_four_2005[20]). This situation presents a challenge for any researcher applying statistical methods to archaeological chronologies with artifacts, as one must make _ad hoc_ and often awkward decisions about dates of types, resulting in what @lavan_checklist_2021[15] called "abonimable dates."
+Archaeological chronology-building centers on two types of dates, absolute and relative. Absolute dates are associated with a calendrical date (even approximate), and relative refer to events only known via relationships to others (before/after). Relative events are then associated to absolute via contextual or logical associations. E.g, a _terminus post quem_ (_t.p.q._) of a datable coin found in a soil deposit ensures the deposit came after the production of the coin. Applying probabilistic methods to chronology has traditionally centered on radiocarbon dating, but the broader goal of extending formal methods to all aspects of chronology necessitate a way to address artifact typologies. Artifacts largely persist in being dated by qualitative judgment (e.g, "around the start of the 4th century BCE"). Moreover, typologies are in a constant state of revision and adjustment, such that the dates of types may differ according to authority. The reasons why a type is dated the way it is (i.e., which sites, deposits, and comparisons inform its date) can also become opaque, such that investigators inherently resist chronological revisions in order to assess the bibliography on any one type (on this conservatism, see @rotroff_four_2005[20]). This situation presents a challenge for any researcher applying statistical methods to archaeological chronologies with artifacts, as one must make _ad hoc_ and often awkward decisions about dates of types, resulting in what @lavan_checklist_2021[15] called "abonimable dates."
 
[…]
