### 2025-07-09 696d85d
Update paper.md
