+++ inst/REFERENCES.bib
@@ -299,3 +299,14 @@
 	author = {Lancel, S. and Morel, J.-P. and Thuillier, J.-P.},
 	year = {1982},
 }
+
+
+@article{eddelbuettel_extending_2018,
+	title = {Extending {R} with {C}++: {A} {Brief} {Introduction} to {Rcpp}},
+	volume = {72},
+	doi = {10.1080/00031305.2017.1375990},
+	journal = {The American Statistician},
+	author = {Eddelbuettel, D. and Balamuta, J.J.},
+	year = {2018},
+	pages = {28--36},
+}
