+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: ../inst/REFERENCES.bib
 
 Archaeological chronology-building centers on two types of dates, absolute and relative. Absolute dates are associated with a calendrical date (even approximate), and relative refer to events only known via relationships to others (before/after). Relative events are then associated to absolute via contextual or logical associations. E.g, a _terminus post quem_ (_t.p.q._) of a datable coin found in a soil deposit ensures the deposit came after the production of the coin. Applying probabilistic methods to chronology has traditionally centered on radiocarbon dating, but the broader goal of extending formal methods to all aspects of chronology necessitate a way to address artifact typologies. Artifacts largely persist in being dated by qualitative judgment (e.g, "around the start of the 4th century BCE"). Moreover, typologies are in a constant state of revision and adjustment, such that the dates of types may differ according to authority. The reasons why a type is dated the way it is (i.e., which sites, deposits, and comparisons inform its date) can also become opaque, such that investigators inherently resist chronological revisions in order to assess its bibliography (on this conservatism, see @rotroff_four_2005[20]). This situation presents a challenge for any researcher applying statistical methods to datable artifacts, as one must make _ad hoc_ and often awkward decisions, resulting in what @lavan_checklist_2021[15] has called "abonimable dates."
 
-The `R` package `eratosthenes` (named after Eratosthenes of Cyrene, author of the _Chronographiai_) provides functions for chronology-building, above all to bring artifact dating within the scope of formal mathematical estimation. Hence, an investigator can obtain separate probability density functions (p.d.f.) for the production, use, and deposition of an artifact type, in addition to marginal p.d.f.s for all relative sequential events (howsover determined) and absolute constraints (howsoever defined). It uses Gibbs sampling, by now a conventional Markov Chain Monte Carlo method in archaeological chronology [@geman_stochastic_1984; @buck_bayesian_1996]. Using `Rcpp`, [@eddelbuettel_extending_2018], `eratosthenes` performs a two-step Gibbs routine, the first a preliminiary sampler to select a starting date, and a second main sampler that uses consistent batch means (CBM) as a stopping rule, given that convergence in distribution is assured [@jones_fixed-width_2006; @flegal_markov_2008]. Full reporting on the Monte Carlo standard errors (MCSE) is provided, giving an error in +/- years for each marginal density. Finally, `eratosthenes` provides functions for assessing the level of dependence of events upon each other within the joint conditional. Changes in dates brought about by any alteration in the structure of a chronology can therefore be readily evaluted. In sum, `eratosthenes` allows for expedient revision of chronologies, transparency in the definition of the full joint conditional, and statistics on the certainty of estimates via MCSE. 
[…]
