### 2025-07-26 #8559 @scollinselliott
Potential reviewers:
benmarwick, andrewcparnell, izaromanowska, sebastien-plutniak, nevrome, nfrerebeau, joeroe
