### 2025-08-31 #8559 @arfon
@scollinselliott – apologies for the slow response on this paper. I'm just familiarizing myself with this codebase. I don't see any tests for this software (automated or otherwise). Am I missing something?
