+++ DESCRIPTION
@@ -21,5 +21,7 @@ LinkingTo:
     Rcpp
 Suggests: 
     knitr,
-    rmarkdown
+    rmarkdown,
+    testthat (>= 3.0.0)
 VignetteBuilder: knitr
+Config/testthat/edition: 3
