### 2025-09-19 #8559 @arfon
@benmarwick @nevrome @izaromanowska @joeroe :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _eratosthenes: Synchronizing archaeological chronologies with special focus on artifact types_

Many thanks
Arfon
