+++ README.md
@@ -4,6 +4,7 @@
 # <img src="man/figures/logo.png" align="right" width="150px"/> eratosthenes: Archaeological Synchronism
 
 <!-- badges: start -->
+<a href="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020"><img src="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020/status.svg"></a>
 <!-- badges: end -->
 
 The `R` package `eratosthenes` aims to provide a general, flexible
