### 2025-10-07 #8559 @benmarwick
Thanks for the invite, I'm flat out doing [reproducibility reviews for JAS](https://www.sciencedirect.com/journal/journal-of-archaeological-science/about/reproducibility-at-journal-of-archaeological-science/reproducibility-at-jas) at the moment. So I regret I cannot take this on currently.

I did have a quick look at the repo and do have a suggestion that might help to interest potential reviewers (and attract users). I recommend the readme include a practical example of how the package can be used to answer a specific archaeological question. Currently it all looks very abstract and statistical. I expect most archaeologists will look at it and be unsure how you expect them to use your package. If you could give more concrete guidance in the readme, I believe that would help to show the value of the package to the discipline.
