### 2025-10-07 1ffe06b
tutorial in readme
