+++ inst/REFERENCES.bib
@@ -17,7 +17,7 @@
 	pages = {725--757},
 }
 
-@Article{schmid_c14bazAAR_2019,
+@article{schmid_c14bazAAR_2019,
   title = {{c14bazAAR}: An {R} package for downloading and preparing {C14} dates from different source databases},
   author = {Schmid, C. and Seidensticker, D. and Hinz, M.},
   journal = {Journal of Open Source Software},
@@ -310,3 +310,32 @@
 	year = {2018},
 	pages = {28--36},
 }
+
+
+@article{tyers_roman_1996,
+    title = {Roman amphoras in Britain},
+	volume = {1},
+	doi = {10.11141/ia.1.6},
+	journal = {Internet Archaeology},
+	author = {Tyers, P.},
+	year = {1996},
+	}
+
+
+@article{southampton_roman_2014,
+    title = {Roman Amphorae: a digital resource [data-set]},
+	volume = {1},
+	doi = {10.5284/1028192},
+	journal = {Internet Archaeology},
+	author = {University of Southampton},
+	year = {2014},
+	}
+
+
+@book{loughton_arverni_2014,
+    address = {Oxford},
+    title = {The Arverni and Roman Wine},
+	publisher = {Archaeopress},
+	year = {2014},
+	}
+
