+++ README.md
@@ -4,6 +4,7 @@
 # <img src="man/figures/logo.png" align="right" width="150px"/> eratosthenes: Archaeological Synchronism
 
 <!-- badges: start -->
+
 <a href="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020"><img src="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020/status.svg"></a>
 <!-- badges: end -->
 
@@ -79,6 +80,320 @@ install the package in the `R` command line with `devtools`:
 library(devtools)
 install_github("scollinselliott/eratosthenes", dependencies = TRUE, build_vignettes = TRUE) 
 ```
+
+## Tutorial: Archaeological Example
+
+The Dressel 1B type of amphora (a ceramic shipping container typical of
+the ancient Mediterranean) is a ceramic type defined on the basis of
+morphology, comprising a tall, long-necked, two-handled vessel with a
+concave collared rim and and sharply defined, angular shoulder. To give
+just three chronological summaries of this type, the Dressel 1B type was
+produced/used/deposited:
+
+- in the “last quarter of the second until the last decade of the first
+  century BC” (Southampton 2014);
+- “shortly before the middle of the 1st century BC” and “until c. 10 BC”
+  (Tyers 1996, 2.2);
+- at the earliest “during *c*.100-80 BC”, disappearing “by 30 BC” (*The
+  Arverni and Roman Wine* 2014, 43).
+
+Such determinations are the result of many comparisons of stratigraphic
+and single contexts where the Dressel 1B type has appeared in
+conjunction with absolute constraints (e.g., coinage, datable stamps,
+historical events in the occupation/destruction of sites). It is evident
+in the above statements, too, that there will be greater or lesser
+chronological discrepancies among investigators.
+
+In order to obtain a probability density for any aspect of dating the
+Dressel 1B (that is, its production, use, and/or deposition),
+`eratosthenes` avoids the awkward synthesis of this chronological
+information by estimating dates directly from archaeological record
+itself, in the information provided by contexts (in
+**[sequences](#sequences)**, howsoever established),
+**[finds](#finds)**, which are contained in those contexts, and
+**[absolute constraints](#absolute-constraints)** for those contexts, as
+inputs. In other words, it formalizes the logic of dating finds which is
+already in practice, yielding a probability density instead of a
+qualitative appraisal of the object’s dates of production, use, and
+deposition.
+
+The following aims to be a concrete tutorial that gives step-by-step
+instructions to obtain probability density functions on the production,
+use, and deposition of the Dressel 1B type using `eratosthenes`. Further
+information on the data for this tutorial is found
+**[here](https://volweb.utk.edu/~scolli46/eratosthenes/eda20250628.html)**.
+
+### Creating a Sequences Object
+
+First, sequences are given as vectors, with contexts entered as elements
+earlier to later, from left to right. The sequence `seq1` below gives a
+sequence of just two depositional contexts, from the archaeological site
+of Rirha in Morocco, `Rirha US 5182` and `Rirha US 5154`, with
+`Rirha US 5182` being the earlier of the two:
+
+``` r
+library(eratosthenes)
+seq1 <- c("Rirha US 5182", "Rirha US 5154") 
+```
+
+Multiple sequences can, and typically will, be given. To define seven
+more sequences for this tutorial, we have additional deposits from the
+site of Carthage (Byrsa Hill) and also several shipwrecks ([further
+information
+here](https://volweb.utk.edu/~scolli46/eratosthenes/eda20250628.html)):
+
+``` r
[…]
