+++ README.Rmd
@@ -17,6 +17,7 @@ knitr::opts_chunk$set(
 # <img src="man/figures/logo.png" align="right" width="150px"/> eratosthenes: Archaeological Synchronism
 
 <!-- badges: start -->
+<a href="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020"><img src="https://joss.theoj.org/papers/75ab124a2cbb3b7125a9458650544020/status.svg"></a>
 <!-- badges: end -->
 
 The `R` package `eratosthenes` aims to provide a general, flexible toolkit for archaeological chronology-building by incorporating, computationally, all relevant sources of information on uncertain archaeological or historical dates. Archaeological dates are subject to relational conditions (via seriation or stratigraphic relationships) and absolute constraints (such as radiocarbon dates, datable artifacts, or other known historical events, as _termini post_ or _ante quos_), which prompt the use of a joint conditional probability density to convey those relationships. The date of any one event can then be marginalized from that full, joint conditional distribution. `Rcpp` is used for faster sampling [@eddelbuettel_extending_2018].
@@ -42,6 +43,192 @@ To obtain the current development version of `eratosthenes` from GitHub, install
 library(devtools)
 install_github("scollinselliott/eratosthenes", dependencies = TRUE, build_vignettes = TRUE) 
 ```
+
+## Tutorial: Archaeological Example
+
+The Dressel 1B type of amphora (a ceramic shipping container typical of the ancient Mediterranean) is a ceramic type defined on the basis of morphology, comprising a tall, long-necked, two-handled vessel with a concave collared rim and and sharply defined, angular shoulder. To give just three chronological summaries of this type, the Dressel 1B type was produced/used/deposited:
+
+* in the "last quarter of the second until the last decade of the first century BC" [@southampton_roman_2014];
+* "shortly before the middle of the 1st century BC" and "until c. 10 BC" [@tyers_roman_1996, 2.2];
+* at the earliest "during _c_.100-80 BC", disappearing "by 30 BC" [@loughton_arverni_2014, 43].
+
+Such determinations are the result of many comparisons of stratigraphic and single contexts where the Dressel 1B type has appeared in conjunction with absolute constraints (e.g., coinage, datable stamps, historical events in the occupation/destruction of sites).  It is evident in the above statements, too, that there will be greater or lesser chronological discrepancies among investigators.
+
+In order to obtain a probability density for any aspect of dating the Dressel 1B (that is, its production, use, and/or deposition), `eratosthenes` avoids the awkward synthesis of this chronological information by estimating dates directly from archaeological record itself, in the information provided by contexts (in **[sequences](#sequences)**, howsoever established), **[finds](#finds)**, which are contained in those contexts, and **[absolute constraints](#absolute-constraints)** for those contexts, as inputs. In other words, it formalizes the logic of dating finds which is already in practice, yielding a probability density instead of a qualitative appraisal of the object's dates of production, use, and deposition.
+
+The following aims to be a concrete tutorial that gives step-by-step instructions to obtain probability density functions on the production, use, and deposition of the Dressel 1B type using `eratosthenes`. Further information on the data for this tutorial is found **[here](https://volweb.utk.edu/~scolli46/eratosthenes/eda20250628.html)**.
+
+### Creating a Sequences Object
+
[…]
