### 2025-10-07 3a7bf16
tutorial in readme
