+++ inst/REFERENCES.bib
@@ -335,6 +335,7 @@
 @book{loughton_arverni_2014,
     address = {Oxford},
     title = {The Arverni and Roman Wine},
+	author = {Loughton, M.},
 	publisher = {Archaeopress},
 	year = {2014},
 	}
