### 2025-10-07 #8559 @scollinselliott
@benmarwick Thanks for your suggestion! I've updated the README to lead with an Archaeological Example right after the section on Installation.
