### 2025-10-07 bde45fb
tutorial in readme
