+++ README.Rmd
@@ -46,7 +46,7 @@ install_github("scollinselliott/eratosthenes", dependencies = TRUE, build_vignet
 
 ## Tutorial: Archaeological Example
 
-The Dressel 1B type of amphora (a ceramic shipping container typical of the ancient Mediterranean) is a ceramic type defined on the basis of morphology, comprising a tall, long-necked, two-handled vessel with a concave collared rim and and sharply defined, angular shoulder. To give just three chronological summaries of this type, the Dressel 1B type was produced/used/deposited:
+The Dressel 1B type of amphora (a ceramic shipping container typical of the ancient Mediterranean) is a ceramic type defined on the basis of morphology, comprising a tall, long-necked, two-handled vessel with a concave collared rim and sharply defined, angular shoulder. To give just three chronological summaries of this type, the Dressel 1B type was produced/used/deposited:
 
 * in the "last quarter of the second until the last decade of the first century BC" [@southampton_roman_2014];
 * "shortly before the middle of the 1st century BC" and "until c. 10 BC" [@tyers_roman_1996, 2.2];
