### 2025-10-15 #9260 @editorialbot
[REVIEW]: eratosthenes: Synchronizing archaeological chronologies with special focus on artifact types
**Submitting author:** @scollinselliott (<a href="http://orcid.org/0000-0002-5642-6903">Stephen Collins-Elliott</a>)
**Repository:** https://github.com/scollinselliott/eratosthenes
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.91
