### 2026-02-04 #9260 @sebastien-plutniak
First of all, my excuse for the long delay in reviewing this submission.
Hopefully, it is a valuable contribution almost ready for publication as it is.

I only have few suggestions:
* add Community guidelines (as requested by JOSS policy)
* in the 'State of the art' section, consider adding references to:
    *  *periodclust* ([app](https://analytics.huma-num.fr/ModAthom/hclustcompro/), [paper](https://doi.org/10.1007/978-3-030-60104-1_1))
    * *cerardate* ([app](http://analytics.huma-num.fr/ModAthom/cerardat/), paper[paper](https://doi.org/10.1016/j.jas.2011.06.031)
* complete the reference 'rotroff_four_2005' (the requested 'publisher' field is empty)

Thank you @scollinselliott  for this contribution!
