### 2026-02-05 #9260 @scollinselliott
Many thanks for these comments @sebastien-plutniak ! I will start on these changes awaiting @joeroe 's review.
