### 2026-04-14 #9260 @arfon
:wave: @simoncarrignon, @bocinsky, @steko, @CaitlinBuck – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is eratosthenes: Synchronizing archaeological chronologies with special focus on artifact types

Many thanks
Arfon
