### 2026-08-25 #9260 @arfon
Thanks @joeroe and for the updates @scollinselliott. @sebastien-plutniak – it looks like the authors have updated their submission based on your feedback – would you mind taking a quick look to verify these changes and complete the missing review checklist items accordingly?
